Make codeblock "copy to clipboard" accessible
The initial implementation of the "copy to clipboard" button was not
easily keyboard navigable. I had to make some changes to the
implementation to make it keyboard navigable. But otherwise, the
functionality has not changed in a discernable way.

I learned this while attempting to write Capybara tests for the
feature.